requirements.txt
pyodbc
azure-identity
flask